+++ src/fft/mod.rs
@@ -42,11 +42,6 @@ SOFTWARE.
 //! `crate::tests::test_spectrum_and_visualize_sine_waves_50_1000_3777hz` by invoking it
 //! with different features (FFT implementations) enabled.
 
-#[cfg(feature = "microfft-complex")]
-mod microfft_complex;
-#[cfg(feature = "microfft-complex")]
-pub use microfft_complex::*;
-
 #[cfg(feature = "microfft-real")]
 mod microfft_real;
 #[cfg(feature = "microfft-real")]