+++ CHANGELOG.md
@@ -4,9 +4,15 @@
 - dependency updates
 - MSRV is now `1.61.0`
 - `FrequencySpectrum::apply_scaling_fn` now requires a reference to `&mut self`:
-  This is breaking but only for a small percentage of users.
+  This is breaking but only for a small percentage of users. Performance is
+  slightly improved as less heap allocations are required.
 - `FrequencySpectrum` is now `Send` and interior mutability is dropped:
   You can wrap the struct in a `Mutex` or similar types now!
+- `FrequencySpectrum::to_map` doesn't  has the `scaling_fn` parameter anymore.
+  This is breaking but only for a small percentage of users.
+- `FrequencySpectrum::to_mel_map` added for getting the spectrum in the
+  [mel](https://en.wikipedia.org/wiki/Mel_scale) scale.
+-
 - small internal code quality and performance improvements
 
 ## 1.2.6 (2022-07-20)